// src/app/components/types/course.ts
export interface Course {
  id: string;
  title: string;
  instructor: string;
  lessons: number;
  duration: string;
  comments: number;
  rating: number;
  level: string;
  price: string;
  billingType: string;
  image: string;
  studentsAvatars: string[];
  totalStudents: string;
  category: string;
}